chore: remove old osc implementation
